+++ tests/api_client/test_api_gateway.py
@@ -1,5 +1,3 @@
-# This module is a copy paste of test_api_gateway.py
-
 from http import HTTPStatus
 from unittest import mock
 
@@ -9,33 +7,33 @@ from clean_python import DoesNotExist
 from clean_python import Json
 from clean_python import Mapper
 from clean_python.api_client import ApiException
-from clean_python.api_client import SyncApiGateway
-from clean_python.api_client import SyncApiProvider
+from clean_python.api_client import ApiGateway
+from clean_python.api_client import ApiProvider
 
 
-class TstSyncApiGateway(SyncApiGateway, path="foo/{id}"):
+class TstApiGateway(ApiGateway, path="foo/{id}"):
     pass
 
 
 @pytest.fixture
 def api_provider():
-    return mock.MagicMock(spec_set=SyncApiProvider)
+    return mock.MagicMock(spec_set=ApiProvider)
 
 
 @pytest.fixture
-def api_gateway(api_provider) -> SyncApiGateway:
-    return TstSyncApiGateway(api_provider)
+def api_gateway(api_provider) -> ApiGateway:
+    return TstApiGateway(api_provider)
 
 
-def test_get(api_gateway: SyncApiGateway):
-    actual = api_gateway.get(14)
+async def test_get(api_gateway: ApiGateway):
+    actual = await api_gateway.get(14)
 
     api_gateway.provider.request.assert_called_once_with("GET", "foo/14")
     assert actual is api_gateway.provider.request.return_value
 
 
-def test_add(api_gateway: SyncApiGateway):
-    actual = api_gateway.add({"foo": 2})
+async def test_add(api_gateway: ApiGateway):
+    actual = await api_gateway.add({"foo": 2})
 
     api_gateway.provider.request.assert_called_once_with(
         "POST", "foo/", json={"foo": 2}
@@ -43,23 +41,23 @@ def test_add(api_gateway: SyncApiGateway):
     assert actual is api_gateway.provider.request.return_value
 
 
-def test_remove(api_gateway: SyncApiGateway):
-    actual = api_gateway.remove(2)
+async def test_remove(api_gateway: ApiGateway):
+    actual = await api_gateway.remove(2)
 
     api_gateway.provider.request.assert_called_once_with("DELETE", "foo/2")
     assert actual is True
 
 
-def test_remove_does_not_exist(api_gateway: SyncApiGateway):
+async def test_remove_does_not_exist(api_gateway: ApiGateway):
     api_gateway.provider.request.side_effect = ApiException(
         {}, status=HTTPStatus.NOT_FOUND
     )
-    actual = api_gateway.remove(2)
+    actual = await api_gateway.remove(2)
     assert actual is False
 
 
-def test_update(api_gateway: SyncApiGateway):
-    actual = api_gateway.update({"id": 2, "foo": "bar"})
+async def test_update(api_gateway: ApiGateway):
+    actual = await api_gateway.update({"id": 2, "foo": "bar"})
 
     api_gateway.provider.request.assert_called_once_with(
         "PATCH", "foo/2", json={"foo": "bar"}
@@ -67,19 +65,19 @@ def test_update(api_gateway: SyncApiGateway):
     assert actual is api_gateway.provider.request.return_value
 
 
-def test_update_no_id(api_gateway: SyncApiGateway):
+async def test_update_no_id(api_gateway: ApiGateway):
     with pytest.raises(DoesNotExist):
-        api_gateway.update({"foo": "bar"})
+        await api_gateway.update({"foo": "bar"})
 
     assert not api_gateway.provider.request.called
 
 
-def test_update_does_not_exist(api_gateway: SyncApiGateway):
+async def test_update_does_not_exist(api_gateway: ApiGateway):
     api_gateway.provider.request.side_effect = ApiException(
         {}, status=HTTPStatus.NOT_FOUND
     )
     with pytest.raises(DoesNotExist):
-        api_gateway.update({"id": 2, "foo": "bar"})
+        await api_gateway.update({"id": 2, "foo": "bar"})
 
 
 class TstMapper(Mapper):
@@ -95,35 +93,35 @@ class TstMapper(Mapper):
         return {"id": external["id"], "name": external["name"].lower()}
 
 
-class TstMappedSyncApiGateway(SyncApiGateway, path="foo/{id}"):
+class TstMappedApiGateway(ApiGateway, path="foo/{id}"):
     mapper = TstMapper()
 
 
 @pytest.fixture
-def mapped_api_gateway(api_provider) -> SyncApiGateway:
-    return TstMappedSyncApiGateway(api_provider)
+def mapped_api_gateway(api_provider) -> ApiGateway:
+    return TstMappedApiGateway(api_provider)
 
 
-def test_get_with_mapper(mapped_api_gateway: SyncApiGateway):
+async def test_get_with_mapper(mapped_api_gateway: ApiGateway):
     mapped_api_gateway.provider.request.return_value = {"id": 14, "name": "FOO"}
 
-    assert mapped_api_gateway.get(14) == {"id": 14, "name": "foo"}
+    assert await mapped_api_gateway.get(14) == {"id": 14, "name": "foo"}
 
 
-def test_add_with_mapper(mapped_api_gateway: SyncApiGateway):
+async def test_add_with_mapper(mapped_api_gateway: ApiGateway):
     mapped_api_gateway.provider.request.return_value = {"id": 3, "name": "FOO"}
 
-    assert mapped_api_gateway.add({"name": "foo"}) == {"id": 3, "name": "foo"}
+    assert await mapped_api_gateway.add({"name": "foo"}) == {"id": 3, "name": "foo"}
 
     mapped_api_gateway.provider.request.assert_called_once_with(
         "POST", "foo/", json={"name": "FOO"}
     )
 
 
-def test_update_with_mapper(mapped_api_gateway: SyncApiGateway):
+async def test_update_with_mapper(mapped_api_gateway: ApiGateway):
     mapped_api_gateway.provider.request.return_value = {"id": 2, "name": "BAR"}
 
-    assert mapped_api_gateway.update({"id": 2, "name": "bar"}) == {
+    assert await mapped_api_gateway.update({"id": 2, "name": "bar"}) == {
         "id": 2,
         "name": "bar",
     }